backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
httpx
pydantic
python-multipart